requirements.txt
Flask>=2.2.2
flask-restful
flask-cors
psycopg2-binary